Set nodefaultdata flag to workaround RIL bug
We unfortunately need to set the "nodefaultdata" flag in
ro.telephony.ril.config to disable setDefaultDataSubId
in "SubscriptionController.java" which is causing the
RIL to crash for some odd reason.

A permanent solution is being worked on but far from ready yet.
Until the permanent solution is in place this flag will remain set.

Change-Id: I708bb61e127c4a48fafa50b74311a7539c55a3f0